amplitude: Clean up scripts and executables.
Create a new directory in amplitude and move all the shell and root scripts
there, as they are supposed to be replaced by python scripts. Also move the
evtTreeHelper files there, as they too should be recreated in a new form. Adapt
the CMakeLists.txt so that compilation works.
