+++ decayAmplitude/test/CMakeLists.txt
@@ -46,13 +46,11 @@ include_directories(
 
 
 # executables
-make_executable(testParticle                testParticle.cc                  "${RPWA_DECAYAMPLITUDE_LIB}" "${RPWA_PARTICLEDATA_LIB}")
+make_executable(testParticle                testParticle.cc                  "${RPWA_DECAYAMPLITUDE_LIB}")
 make_executable(testDecayTopology           testDecayTopology.cc             "${RPWA_DECAYAMPLITUDE_LIB}" "${RPWA_UTILITIES_LIB}")
 make_executable(testWaveDescription         testWaveDescription.cc           "${RPWA_DECAYAMPLITUDE_LIB}" "${RPWA_UTILITIES_LIB}")
 make_executable(testWaveSetGenerator        testWaveSetGenerator.cc          "${RPWA_DECAYAMPLITUDE_LIB}")
-make_executable(testAmplitude               testAmplitude.cc                 "${RPWA_DECAYAMPLITUDE_LIB}" "${RPWA_UTILITIES_LIB}")
 make_executable(testAmplitudeTree           testAmplitudeTree.cc             "${RPWA_DECAYAMPLITUDE_LIB}" "${RPWA_UTILITIES_LIB}")
-make_executable(testIsospinSym              testIsospinSym.cc                "${RPWA_DECAYAMPLITUDE_LIB}" "${RPWA_UTILITIES_LIB}")
 make_executable(testIntegral                testIntegral.cc                  "${RPWA_DECAYAMPLITUDE_LIB}")
 make_executable(testPeakShapeInstableIsobar testPeakShapeInstableIsobar.cc   "${RPWA_DECAYAMPLITUDE_LIB}" "${RPWA_GENERATORS_LIB}")
 make_executable(testPeakShapeInstableIsobar2 testPeakShapeInstableIsobar2.cc "${RPWA_DECAYAMPLITUDE_LIB}")
